Backend/emails: Consolidate render_email
