+++ app/backend/src/couchers/email/queuing.py
@@ -6,8 +6,8 @@ from sqlalchemy.orm.session import Session
 
 from couchers.config import config
 from couchers.context import CouchersContext
-from couchers.email.emails import EmailBase
-from couchers.email.rendering import EmailFooter, render_html_body, render_plaintext_body
+from couchers.email.blocks import EmailBase, EmailFooter
+from couchers.email.rendering import render_email
 from couchers.i18n import LocalizationContext
 from couchers.jobs.enqueue import queue_job
 from couchers.metrics import emails_counter
@@ -51,16 +51,8 @@ def queue_userless_email(
     if not context.get_boolean_value("notification_translations_enabled", default=False):
         loc_context = LocalizationContext(locale="en", timezone=loc_context.timezone)
 
-    subject = email.get_subject_line(loc_context)
-    preview = email.get_preview_line(loc_context)
-    body_blocks = email.get_body_blocks(loc_context)
-
     footer = EmailFooter(timezone_name=loc_context.localized_timezone, copyright_year=now().year, unsubscribe_info=None)
-
-    body_plaintext = render_plaintext_body(blocks=body_blocks, footer=footer, loc_context=loc_context)
-    body_html = render_html_body(
-        subject=subject, preview=preview, blocks=body_blocks, footer=footer, loc_context=loc_context
-    )
+    rendered = render_email(email, footer, loc_context)
 
     queue_email(
         session,
@@ -68,9 +60,9 @@ def queue_userless_email(
             sender_name=config.NOTIFICATION_EMAIL_SENDER,
             sender_email=config.NOTIFICATION_EMAIL_ADDRESS,
             recipient=recipient,
-            subject=config.NOTIFICATION_PREFIX + subject,
-            plain=body_plaintext,
-            html=body_html,
+            subject=config.NOTIFICATION_PREFIX + rendered.subject,
+            plain=rendered.body_plaintext,
+            html=rendered.body_html,
             source_data=f"{source_data_header}; version={config.VERSION}",
         ),
     )
